1) Removed and suppressed format mask warnings. 2) Changed initial value of _networkvailable. 3) Changed timeFormat methods so warnings removed. 4) Changed some return types to match updated sdk. 5) Added typecasts to MD5 code. 6) Updated ELog macro to match weSpot.

+++ ARLearn/ARLAudioPlayer.m
@@ -225,11 +225,11 @@
     float minutes = floor(lroundf([value floatValue])/60);
     float seconds = lroundf([value floatValue]) - (minutes * 60);
     
-    int roundedSeconds = lroundf(seconds);
-    int roundedMinutes = lroundf(minutes);
+    long roundedSeconds = lroundf(seconds);
+    long roundedMinutes = lroundf(minutes);
     
     NSString *time = [[NSString alloc]
-                      initWithFormat:@"%d:%02d",
+                      initWithFormat:@"%ld:%02ld",
                       roundedMinutes, roundedSeconds];
     return time;
 }
